Convert lib/shorteners to ts
